requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
boto3==1.14.23
botocore==1.17.23
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.0.8
django-bootstrap4==2.2.0
django-heroku==0.3.1
django-mapbox-location-field==1.6.1
django-registration==3.1
django-storages==1.9.1
docutils==0.15.2
folium==0.11.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.0
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
zipp==3.1.0
